+++ packages/ui-core/src/components/Carousel/Carousel.tsx
@@ -22,6 +22,12 @@ interface ICarouselOptionsResponsive {
     settings: Pick<ICarouselOptions, 'slidesToShow' | 'arrows'>;
 }
 
+interface ICarouselClasses {
+    root?: string;
+    leftArrow?: string;
+    rightArrow?: string;
+}
+
 export interface ICarouselProps {
     className?: string;
     options: ICarouselOptions;
@@ -29,6 +35,7 @@ export interface ICarouselProps {
     arrowColor?: string;
     /** Padding between slides */
     disablePaddingBetweenSlides?: boolean;
+    classes?: ICarouselClasses;
     children: any;
     onClickNext?: () => void;
     onClickPrev?: () => void;
@@ -60,6 +67,11 @@ class Carousel extends React.Component<ICarouselProps, ICarouselState> {
                 PropTypes.array,
             ])
         ),
+        classes: PropTypes.shape({
+            root: PropTypes.string,
+            leftArrow: PropTypes.string,
+            rightArrow: PropTypes.string,
+        }),
         theme: PropTypes.oneOf(['default', 'landing', 'showcase', 'lk']),
         arrowColor: PropTypes.oneOf(['white']),
         disablePaddingBetweenSlides: PropTypes.bool,
@@ -250,18 +262,20 @@ class Carousel extends React.Component<ICarouselProps, ICarouselState> {
     }
 
     renderArrows() {
-        const { theme, arrowColor } = this.props;
+        const { theme, arrowColor, classes = {} } = this.props;
         const { isPrevActive, isNextActive } = this.state;
+        const modPrevArrow = { 'arrow-prev': true, disabled: !isPrevActive, fill: arrowColor };
+        const modNextArrow = { 'arrow-next': true, disabled: !isNextActive, fill: arrowColor };
 
         return (
             <div className={cn('arrows', { theme: theme })}>
                 <CarouselArrow
-                    className={cn('arrow', { 'arrow-prev': true, disabled: !isPrevActive, fill: arrowColor })}
+                    className={cn('arrow', modPrevArrow, classes.leftArrow)}
                     onClick={this.handleClickPrev}
                     theme={theme}
                 />
                 <CarouselArrow
-                    className={cn('arrow', { 'arrow-next': true, disabled: !isNextActive, fill: arrowColor })}
+                    className={cn('arrow', modNextArrow, classes.rightArrow)}
                     onClick={this.handleClickNext}
                     theme={theme}
                 />
@@ -270,14 +284,15 @@ class Carousel extends React.Component<ICarouselProps, ICarouselState> {
     }
 
     render() {
-        const { className, options, theme, children, onBeforeChange, disablePaddingBetweenSlides } = this.props;
+        const { options, theme, children, onBeforeChange, disablePaddingBetweenSlides } = this.props;
         const { isArrows } = this.state;
         const { arrows, ...carouselOptions } = options;
+        const { className = '', classes = { root: '' } } = this.props;
+        const modClasses = { theme, 'no-padding-between-slides': disablePaddingBetweenSlides };
+        const propsClassName = `${className} ${classes.root}`.trim();
 
         return (
-            <div
-                className={cn('', { theme, 'no-padding-between-slides': disablePaddingBetweenSlides }, className)}
-            >
+            <div className={cn('', modClasses, propsClassName)}>
                 {isArrows && this.renderArrows()}
                 <Slider
                     {...carouselOptions}