+++ packages/ui-core/CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 ## Unreleased
-
-## Unreleased
 - **Carousel**
-    - delete margin beetween slides in landign theme
+    - add classes props for flexible styling of the component
+
+## Release 1.3.3
+- **Carousel**
+    - remove margin beetween slides in landing theme
 
 ## Release 1.3.2
 - **Carousel**